+++ packages/analytics-js-plugins/jest.config.mjs
@@ -3,6 +3,7 @@ export default {
   displayName: '@rudderstack/analytics-js-plugins',
   preset: '../../jest.preset.js',
   rootDir: './',
+  coveragePathIgnorePatterns: ['src/errorReporting/event'],
   coverageThreshold: {
     global: {
       branches: 0,